Separate Logger context provider and useLogger hook for react-refresh
