+++ packages/devtools/src/components/ChatMessage/ChatMessage.tsx
@@ -19,7 +19,7 @@ import { ChatContext } from '../../stores/ChatStore';
 import useSparkApi from '../../hooks/useSparkApi';
 import AttachmentsContainer from '../AttachmentsContainer/AttachmentsContainer';
 import { AttachmentType } from '../../types/Attachment';
-import { useLogger } from '../../contexts/LoggerContext';
+import useLogger from '../../hooks/useLogger';
 
 interface ChatMessageProps {
   content: string;
@@ -47,16 +47,15 @@ const ChatMessage: FC<ChatMessageProps> = ({
     (value.body?.contentType === 'text' && value.body?.content) || ''
   );
   const [reactions, setReactions] = useState<MessageReaction[]>(value.reactions || []);
+  const [reactionSender, setReactionSender] = useState<MessageUser | undefined>(undefined);
   const hasAttachments = value.attachments && value.attachments.length > 0;
-
-  let reactionSender: MessageUser | undefined;
 
   const handleMessageReaction = async (id: string, newReactionActivity: MessageReaction) => {
     const message = messages[chat.id].find((m) => m.id === id);
 
     if (!message) return;
     const { type, user } = newReactionActivity;
-    reactionSender = user;
+    setReactionSender(user);
     const added: Array<MessageReaction> = [];
     const removed: Array<MessageReaction> = [];
     const reaction = (message.reactions || []).find(
