+++ packages/devtools/src/contexts/LoggerContext.tsx
@@ -1,8 +1,6 @@
 import { ConsoleLogger, LogLevel } from '@microsoft/spark.common/logging';
-import { createContext, FC, PropsWithChildren, useContext } from 'react';
-
-const logger = new ConsoleLogger('@spark/devtools');
-export const LoggerContext = createContext<ConsoleLogger>(logger);
+import { FC, PropsWithChildren } from 'react';
+import { LoggerContext } from './LoggerContextCore';
 
 export const LoggerProvider: FC<PropsWithChildren<{ level?: LogLevel }>> = ({
   children,
@@ -12,11 +10,4 @@ export const LoggerProvider: FC<PropsWithChildren<{ level?: LogLevel }>> = ({
   return <LoggerContext.Provider value={contextLogger}>{children}</LoggerContext.Provider>;
 };
 
-export const useLogger = () => {
-  const context = useContext(LoggerContext);
-  if (!context) {
-    logger.error('Logger context not found - useLogger must be used within LoggerProvider');
-    throw new Error('useLogger must be used within a LoggerProvider');
-  }
-  return context;
-};
+export { LoggerContext } from './LoggerContextCore';
