+++ packages/devtools/src/hooks/useLogger.ts
@@ -1,0 +1,13 @@
+import { useContext } from 'react';
+import { LoggerContext, logger } from '../contexts/LoggerContextCore';
+
+const useLogger = () => {
+  const context = useContext(LoggerContext);
+  if (!context) {
+    logger.error('Logger context not found - useLogger must be used within LoggerProvider');
+    throw new Error('useLogger must be used within a LoggerProvider');
+  }
+  return context;
+};
+
+export default useLogger;
