+++ packages/devtools/src/screens/ChatScreen/ChatScreen.tsx
@@ -1,17 +1,17 @@
-import { FC, useContext, useCallback, useState } from 'react';
-import { ChatContext } from '../../stores/ChatStore';
-import { useClasses } from './ChatScreen.styles';
 import { Attachment } from '@microsoft/spark.api';
-import useSparkApi from '../../hooks/useSparkApi';
-import { useDevModeSendMessage } from '../../utils/dev';
+import { FC, useCallback, useContext, useState } from 'react';
 
+import useSparkApi from '../../hooks/useSparkApi';
+import { ChatContext } from '../../stores/ChatStore';
+import { useDevModeSendMessage } from '../../utils/dev';
+import { useClasses } from './ChatScreen.styles';
 import Chat from '../../components/Chat/Chat';
-import ChatMessageContainer from '../../components/ChatMessage/ChatMessageContainer';
 import ChatMessage from '../../components/ChatMessage/ChatMessage';
+import ChatMessageContainer from '../../components/ChatMessage/ChatMessageContainer';
 import ComposeBox from '../../components/ComposeBox/ComposeBox';
 import TypingIndicator from '../../components/TypingIndicator/TypingIndicator';
+import useLogger from '../../hooks/useLogger';
 import { useScreensClasses } from '../Screens.styles';
-import { useLogger } from '../../contexts/LoggerContext';
 
 const MAX_HISTORY = 5;
 
@@ -41,7 +41,7 @@ const ChatScreen: FC<ChatScreenProps> = ({ isConnected }) => {
         childLog.error('Error sending message:', err);
       }
     },
-    [sparkApi, chat?.id]
+    [sparkApi, chat?.id, childLog]
   );
 
   const handleMessageSent = useCallback((message: string) => {
